Separate dotnet and node functions into modules
