+++ src/versions/package_version.ts
@@ -1,7 +1,5 @@
-import * as fs from 'fs';
-import * as parser from 'xml2json';
-import * as path from 'path';
-import { sh } from '../cli/shell';
+import { getPackageVersionDotnet, setPackageVersionDotnet } from './package_version/dotnet';
+import { getPackageVersionNode, setPackageVersionNode } from './package_version/node';
 
 const PACKAGE_MODE_DOTNET = 'dotnet';
 const PACKAGE_MODE_NODE = 'node';
@@ -26,20 +24,6 @@ export function getPackageVersionTag(mode): string {
   return `v${getPackageVersion(mode)}`;
 }
 
-function getPackageVersionDotnet(): string {
-  const pathToCsproj = getCsprojPath();
-  const version = JSON.parse(getJsonFromFile(pathToCsproj)).Project.PropertyGroup.Version;
-
-  return version;
-}
-
-function getPackageVersionNode(): string {
-  const rawdata = fs.readFileSync('package.json').toString();
-  const packageJson = JSON.parse(rawdata);
-
-  return packageJson.version;
-}
-
 export function setPackageVersion(mode: string, version: string): void {
   if (mode === PACKAGE_MODE_DOTNET) {
     setPackageVersionDotnet(version);
@@ -55,73 +39,4 @@ function getMajorVersion(version: string): string {
   const majorVersion: string = regexResult[1];
 
   return majorVersion;
-}
-
-function getJsonFromFile(filePath: string): string {
-  const csproj = fs.readFileSync(filePath, { encoding: 'utf8' });
-
-  const jsonString = parser.toJson(csproj);
-
-  return jsonString;
-}
-
-function setPackageVersionNode(version: string): void {
-  sh(`npm version ${version} --no-git-tag-version`);
-}
-
-function setPackageVersionDotnet(newVersion: string): void {
-  const currentVersion = getPackageVersionDotnet();
-  if (currentVersion == null) {
-    throw new Error('Unexpected value: `currentVersion` should not be null here.');
-  }
-
-  const pathToCsproj = getCsprojPath();
-  const csProj = fs.readFileSync(pathToCsproj, { encoding: 'utf8' });
-  const csProjWithNewVersion = csProj.replace(
-    `<Version>${currentVersion}</Version>`,
-    `<Version>${newVersion}</Version>`
-  );
-
-  fs.writeFileSync(pathToCsproj, csProjWithNewVersion);
-}
-
-function getCsprojPath(): string {
-  if (process.platform === 'win32') {
-    return getCsprojPathOnWindows();
-  }
-
-  return getCsprojPathOnNonWindows();
-}
-
-function getCsprojPathOnWindows(): string {
-  const result = sh('where /r . *.csproj');
-  const paths = result.split('\n');
-
-  const relativeCsprojPath = findRelativeCsprojPath(paths);
-  return relativeCsprojPath.replace(/\r/g, '');
-}
-
-function getCsprojPathOnNonWindows(): string {
-  const result = sh('find . -print | grep -i .csproj');
-  const paths = result.split('\n');
-
-  return findRelativeCsprojPath(paths);
-}
-
-function findRelativeCsprojPath(paths: Array<string>): string {
-  const filteredPaths = paths.filter((filePath: string) => {
-    // Replace the current working dir, because Windows returns absolute paths when using `where`
-    const relativePathToCsproj = filePath.trim().replace(process.cwd(), '');
-    const parsedPath = path.parse(relativePathToCsproj);
-
-    return (
-      relativePathToCsproj.endsWith('.csproj') && !parsedPath.dir.includes('test') && !parsedPath.dir.includes('tests')
-    );
-  });
-
-  if (filteredPaths.length > 1) {
-    throw new Error(`More than one .csproj file found: ${filteredPaths}`);
-  }
-
-  return filteredPaths[0];
 }
